[polaris.shopify.com] Fix figma usage value (#6497)
* Fix incorrect value in figma usage column

* Add changeset

+++ polaris.shopify.com/src/components/TokenList/TokenList.tsx
@@ -87,7 +87,7 @@ function getFigmaUsageForToken(
   } else if (name.startsWith("shadow")) {
     usage = "Use shadow styles from UI kit";
   } else if (name.includes("breakpoint")) {
-    const artboardWidth = parseInt(value) * REM;
+    const artboardWidth = parseFloat(value) * REM;
     if (artboardWidth > 0) {
       usage = `Set frame width to ${artboardWidth}+ pixels`;
     }
